examples/python/greenlet_consumer/requirements.txt
simplejson
gevent
pyzmq